feat(runtime): extract /meta and /data — ADR-0076 D11 step ③ PR-10, terminal cut (#2462)
The last two if-chain domains leave the dispatcher:
- domains/meta.ts: the metadata surface (ADR-0033 draft-aware protocol
  paths, org-scoped reads, requireAuth gate via deps.isAuthRequired)
  with its exclusive slimDocList (ADR-0046) riding along.
- domains/data.ts: CRUD/query over actionExec.callData; the
  multi-tenant unresolved-env 428 keys off a new semantic
  deps.isMultiTenantHost() instead of poking this.kernelResolver.

dispatch() now contains ZERO domain branches — 18 domains resolve
through the DomainHandlerRegistry; the gate stages + registry lookup is
all that remains. createHonoApp's catch-all is ready for retirement
(step ① of #2462).

Verified: runtime 649, http-conformance 41, dogfood 351, full build
green.
